requirements.txt
dataset==1.6.2
hikari==2.0.0.dev125
hikari_lightbulb==2.3.5
python-dotenv==1.0.1
Requests==2.31.0
uvloop==0.19.0
